fix & move process_metrics to ranking_metric
